Test framework improvement, Windows version bug fixes
Removed some weak symbols from the test framework by separating
the stubs into a file included in thebuild of the test which actually
use stub symbols.

Some bugfixes in the Windows portion of the arch files

Criticism and suggestions are always welcome!
