+++ src/instr/meson.build
@@ -10,7 +10,7 @@ instr_cfg = custom_target(
 
 rootsim_llvm_lib_dep = [llvm_dep]
 if get_option('b_sanitize') == 'address'
-  rootsim_llvm_lib_dep += meson.get_compiler('c').find_library('asan', 
+  rootsim_llvm_lib_dep += meson.get_compiler('c').find_library('asan',
     static : true)
 endif
 
@@ -20,9 +20,9 @@ rootsim_llvm_lib = shared_library('rootsim-llvm', ['instr_llvm.cpp', instr_cfg],
   cpp_args : log_level_def, include_directories : src_inc_dir, install : true)
 
 clang_arguments = c_optimization_options
-clang_arguments += ['@EXTRA_ARGS@', '-I', meson.source_root() / 'src', 
-  '-I', '@CURRENT_SOURCE_DIR@', '@INPUT@', '-c', '-o', '@OUTPUT@', 
+clang_arguments += ['@EXTRA_ARGS@', '-I', meson.source_root() / 'src',
+  '-I', '@CURRENT_SOURCE_DIR@', '@INPUT@', '-c', '-o', '@OUTPUT@',
   '-Xclang', '-load', '-Xclang', rootsim_llvm_lib.full_path()]
 
-clang_with_plugin = generator(clang_compiler, output : '@BASENAME@.o', 
+clang_with_plugin = generator(clang_compiler, output : '@BASENAME@.o',
   arguments : clang_arguments, depends : rootsim_llvm_lib)
